refactor(CSS)CSS in JS 취지에 맞게끔 리팩토링
