+++ src/container/CategoriesContainer.jsx
@@ -1,14 +1,36 @@
 import { useDispatch, useSelector } from 'react-redux';
 
+import styled from '@emotion/styled';
+
 import {
   selectCategory,
   loadRestaurants,
 } from '../slice';
 
-import MenuItem from '../styles/MenuItem';
-import Menu from '../styles/Menu';
-
 import { get } from '../utils';
+
+const Menu = styled.ul({
+  display: 'flex',
+  margin: 0,
+  padding: 0,
+  listStyle: 'none',
+});
+
+const MenuItem = styled.li(({ active }) => ({
+  marginRight: '1em',
+  '& button': {
+    marginBottom: '1em',
+    padding: '.4em 1em',
+    border: '1px solid #CCCCCC',
+    color: '#333333',
+    background: active ? '#EEEEEE' : 'transparent',
+    cursor: 'pointer',
+    '&:hover': {
+      fontWeight: 'bold',
+      color: '#000000',
+    },
+  },
+}));
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
