+++ src/container/RegionsContainer.jsx
@@ -1,5 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 
+import styled from '@emotion/styled';
+
 import {
   selectRegion,
   loadRestaurants,
@@ -7,8 +9,28 @@ import {
 
 import { get } from '../utils';
 
-import Menu from '../styles/Menu';
-import MenuItem from '../styles/MenuItem';
+const Menu = styled.ul({
+  display: 'flex',
+  margin: 0,
+  padding: 0,
+  listStyle: 'none',
+});
+
+const MenuItem = styled.li(({ active }) => ({
+  marginRight: '1em',
+  '& button': {
+    marginBottom: '1em',
+    padding: '.4em 1em',
+    border: '1px solid #CCC',
+    color: '#333333',
+    background: active ? '#EEE' : 'transparent',
+    cursor: 'pointer',
+    '&:hover': {
+      fontWeight: 'bold',
+      color: '#000',
+    },
+  },
+}));
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
