--- src/styles/Menu.jsx
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-const Menu = styled.ul({
-  display: 'flex',
-  margin: 0,
-  padding: 0,
-  listStyle: 'none',
-});
-
-export default Menu;
